backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
python-multipart==0.0.6
email-validator==2.1.0
aiohttp==3.9.1
transformers>=4.35.0
torch>=2.1.0
pytz==2023.3
redis==5.0.1